VMS-9818: Fix non monotonic timestamps issue with interleaved audio video packets.
<==== Mercurial data ====>
Branch: webcam_plugin

+++ plugins/usb_cam/src/ffmpeg/utils.cpp
@@ -266,7 +266,7 @@ void toFraction(float number, int * outNumerator, int * outDenominator)
     *outDenominator = den;
 }
 
-int suggestChannelLayout(AVCodec *codec)
+uint64_t suggestChannelLayout(AVCodec *codec)
 {
     if (!codec->channel_layouts)
         return AV_CH_LAYOUT_STEREO;
